+++ src/vector/arithmetic/vector_sub_dd.cpp
@@ -8,7 +8,7 @@ using namespace ddavx_core;
 
 void dd_real_vector::sub(dd_real_vector& vec1, dd_real_vector& vec2)
 {
-	if(size() != vec1.size() || size() != vec2.size()){
+	if((long)size() != (long)vec1.size() || (long)size() != (long)vec2.size()){
 		std::cout << "error vecvor size is" << size() << vec1.size() << vec2.size() << std::endl;
 		assert(1);
 	};
@@ -43,7 +43,7 @@ void dd_real_vector::sub(dd_real_vector& vec1, dd_real_vector& vec2)
 
 void dd_real_vector::sub(d_real_vector& vec1, dd_real_vector& vec2)
 {
-	if(size() != (long)vec1.size() || size() != vec2.size()){
+	if((long)size() != (long)vec1.size() || (long)size() != (long)vec2.size()){
 		std::cout << "error vecvor size is" << size() << vec1.size() << vec2.size() << std::endl;
 		assert(1);
 	};
@@ -83,7 +83,7 @@ void dd_real_vector::sub(dd_real_vector& vec1, d_real_vector& vec2)
 
 void dd_real_vector::sub(d_real_vector& vec1, d_real_vector& vec2)
 {
-	if(size() != (long)vec1.size() || size() != (long)vec2.size()){
+	if((long)size() != (long)vec1.size() || (long)size() != (long)vec2.size()){
 		std::cout << "error vecvor size is" << size() << vec1.size() << vec2.size() << std::endl;
 		assert(1);
 	};
